use arbitrary precision integers when converting between LoAT and SMT expressions
# Conflicts:
#	src/lib/expr/numexpression.cpp
#	src/lib/expr/numexpression.hpp
#	src/lib/smt/cvc5context.cpp
#	src/lib/smt/cvc5context.hpp
#	src/lib/smt/exprtosmt.hpp
#	src/lib/smt/smtcontext.hpp
#	src/lib/smt/yicescontext.cpp
#	src/lib/smt/yicescontext.hpp
#	src/lib/smt/z3context.cpp
#	src/lib/smt/z3context.hpp

+++ src/lib/smt/cvc5context.hpp
@@ -38,7 +38,7 @@ public:
     bool isVar(const cvc5::Term &e) const override;
     bool isRationalConstant(const cvc5::Term &e) const override;
     bool isInt(const cvc5::Term &e) const override;
-    long toInt(const cvc5::Term &e) const override;
+    Int toInt(const cvc5::Term &e) const override;
     cvc5::Term lhs(const cvc5::Term &e) const override;
     cvc5::Term rhs(const cvc5::Term &e) const override;
